Refactor GoogleAuth and add tests
