+++ src/GoogleAuth/index.tsx
@@ -1,0 +1,3 @@
+export * from "./GoogleAuth";
+export * from "./parseTorusRedirectParams";
+export * from "./getGoogleCredentials";
