+++ src/GoogleAuth/parseTorusRedirectParams.ts
@@ -1,0 +1,40 @@
+/**
+ * This function parses the deeplink url we get on a successful social auth.
+ *
+ * The parameters are built by
+ * https://github.com/trilitech/umami-web/blob/main/public/auth/v2/redirect.html
+ *
+ * which is essentially a copy of
+ * https://github.com/torusresearch/CustomAuth/blob/master/serviceworker/redirect.html
+ */
+export const parseTorusRedirectParams = (url: string) => {
+  const correctUrl = url.replace("umami://auth/", "");
+  const params = new URLSearchParams(correctUrl);
+  const instanceParams = {
+    insanceId: params.get("instanceId"),
+    verifier: params.get("verifier"),
+    typeOfLogin: params.get("typeOfLogin"),
+    redirectToOpener: params.get("redirectToOpener"),
+  };
+
+  const hashParams = {
+    state: params.get("state"),
+    access_token: params.get("access_token"),
+    token_type: params.get("token_type"),
+    expires_in: params.get("expires_in"),
+    scope: params.get("scope"),
+    id_token: params.get("id_token"),
+    authuser: params.get("authuser"),
+    hd: params.get("hd"),
+    prompt: params.get("prompt"),
+  };
+
+  const data = { instanceParams, hashParams };
+  const result = {
+    channel: params.get("channel"),
+    data: data,
+    error: params.get("error"),
+  };
+
+  return result;
+};
